+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -8,7 +8,7 @@ from demisto_sdk.commands.common.constants import (
     BANG_COMMAND_NAMES, DBOT_SCORES_DICT, DEFAULT_CONTENT_ITEM_FROM_VERSION,
     DEPRECATED_REGEXES, ENDPOINT_COMMAND_NAME, ENDPOINT_FLEXIBLE_REQUIRED_ARGS,
     FEED_REQUIRED_PARAMS, FIRST_FETCH, FIRST_FETCH_PARAM,
-    INCIDENT_FETCH_REQUIRED_PARAMS, INTEGRATION_CATEGORIES, IOC_OUTPUTS_DICT,
+    INCIDENT_FETCH_REQUIRED_PARAMS, IOC_OUTPUTS_DICT,
     MAX_FETCH, MAX_FETCH_PARAM, PACKS_DIR, PACKS_PACK_META_FILE_NAME,
     PYTHON_SUBTYPES, REPUTATION_COMMAND_NAMES, TYPE_PWSH,
     XSOAR_CONTEXT_STANDARD_URL, XSOAR_SUPPORT, MarketplaceVersions)
@@ -26,6 +26,7 @@ from demisto_sdk.commands.common.hook_validations.description import \
     DescriptionValidator
 from demisto_sdk.commands.common.hook_validations.docker import \
     DockerImageValidator
+from demisto_sdk.commands.common import tools
 from demisto_sdk.commands.common.hook_validations.image import ImageValidator
 from demisto_sdk.commands.common.tools import (
     _get_file_id, compare_context_path_in_yml_and_readme,
@@ -351,8 +352,9 @@ class IntegrationValidator(ContentEntityValidator):
         # type: () -> bool
         """Check that the integration category is in the schema."""
         category = self.current_file.get('category', None)
-        if category not in INTEGRATION_CATEGORIES:
-            error_message, error_code = Errors.wrong_category(category)
+        approved_list = tools.get_current_categories()
+        if category not in approved_list:
+            error_message, error_code = Errors.wrong_category(category, approved_list)
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 self.is_valid = False
                 return False