+++ demisto_sdk/commands/common/tests/pack_metadata_validator_test.py
@@ -44,6 +44,8 @@ class TestPackMetadataValidator:
                                           ])
     def test_metadata_validator_valid(self, mocker, metadata):
         mocker.patch.object(tools, 'get_dict_from_file', return_value=({'approved_list': []}, 'json'))
+        mocker.patch('demisto_sdk.commands.common.hook_validations.integration.tools.get_current_categories',
+                     return_value=["Data Enrichment & Threat Intelligence"])
         mocker.patch.object(PackUniqueFilesValidator, '_read_file_content',
                             return_value=TestPackMetadataValidator.read_file(metadata))
         mocker.patch.object(PackUniqueFilesValidator, '_is_pack_file_exists', return_value=True)
@@ -334,3 +336,39 @@ class TestPackMetadataValidator:
         pack, should_pack_be_deprecated = deprecated_pack
         validator = PackUniqueFilesValidator(pack.path)
         assert validator.should_pack_be_deprecated() == should_pack_be_deprecated
+      
+    VALID_CATEGORIES_LIST = ["Endpoint", "File Integrity Management"]
+    
+    @pytest.mark.parametrize("metadata_content, expected_results, valid_list_mock",
+                             [({"categories": ["Endpoint"]}, True, VALID_CATEGORIES_LIST),
+                              ({"categories": ["Analytics & SIEMM"]}, False, VALID_CATEGORIES_LIST),
+                              ({"categories": ["Analytics & SIEMM"]}, False, VALID_CATEGORIES_LIST),
+                              ({"categories": ["Endpoint", "File Integrity Management"]}, False, VALID_CATEGORIES_LIST),
+                              ({"categories": ["Analytics & SIEMM", "random category"]}, False, VALID_CATEGORIES_LIST),
+                              ({"categories": ["Analytics & SIEMM", "Endpoint"]}, False, VALID_CATEGORIES_LIST)])
+    def test_is_categories_field_match_standard(self, mocker, metadata_content, expected_results, valid_list_mock):
+        """
+        Given:
+            - A pack metadata content and a list of approved categories.
+            - case 1: pack metadata content with one valid category and the valid categories list.
+            - case 2: pack metadata content with one invalid category and the valid categories list.
+            - case 3: pack metadata content with two valid categories and the valid categories list.
+            - case 4: pack metadata content with two invalid categories and the valid categories list.
+            - case 5: pack metadata content with one invalid category and one valid category, and the valid categories list.
+
+        When:
+            - running is_categories_field_match_standard function.
+
+        Then:
+            - Ensure that the categories field was validated correctly.
+            - case 1: Should return True.
+            - case 2: Should return False.
+            - case 3: Should return False.
+            - case 4: Should return False.
+            - case 5: Should return False.
+        """
+        mocker.patch('demisto_sdk.commands.common.hook_validations.integration.tools.get_current_categories',
+                     return_value=valid_list_mock)
+        validator = PackUniqueFilesValidator('test')
+        validator.metadata_content = metadata_content
+        assert validator.is_categories_field_match_standard() is expected_results